+++ lib/clipboard_detect.dart
@@ -6,7 +6,29 @@ class ClipboardDetect {
   }
 
   Future<List<String>> detectClipboardPatterns({List<String>? patterns}) {
-    return ClipboardDetectPlatform.instance.detectClipboardPatterns(
+    return ClipboardDetectPlatform.instance.detectClipboardPatterns(patterns: patterns);
+  }
+
+  Future<List<List<String>>> detectClipboardPatternsInItems({
+    List<int>? itemIndexes,
+    List<String>? patterns,
+  }) {
+    return ClipboardDetectPlatform.instance.detectClipboardPatternsInItems(
+      itemIndexes: itemIndexes,
+      patterns: patterns,
+    );
+  }
+
+  Future<Map<String, dynamic>> detectClipboardValues({List<String>? patterns}) {
+    return ClipboardDetectPlatform.instance.detectClipboardValues(patterns: patterns);
+  }
+
+  Future<List<Map<String, dynamic>>> detectClipboardValuesInItems({
+    List<int>? itemIndexes,
+    List<String>? patterns,
+  }) {
+    return ClipboardDetectPlatform.instance.detectClipboardValuesInItems(
+      itemIndexes: itemIndexes,
       patterns: patterns,
     );
   }